+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added the UI pagination to users list
 - Added a permission to sales manager to handle the users (manager and representative)
 - Enabled all the roles on the admin UI interface organization add and edit 
+- Added a new UI component to handle the organizations without a sales manager
 
 ## [1.9.0] - 2022-05-17